going upstairs seperating all the nonsesne
